divided classes into category and product files feature/homework_14_1
