+++ src/utils.py
@@ -1,36 +1,7 @@
 import json
 
-
-class Product:
-    def __init__(self, name: str, description: str, price: float, quantity: int):
-        """Инициализация продукта. Все параметры обязательны."""
-        self.name = name
-        self.description = description
-        self.price = price
-        self.quantity = quantity
-
-    def __repr__(self) -> str:
-        return f"Product('{self.name}', {self.price} руб., {self.quantity} шт.)"
-
-
-class Category:
-    # Атрибуты класса (общие для всех объектов)
-    total_categories = 0  # Общее количество категорий
-    total_products = 0  # Общее количество уникальных товаров
-
-    def __init__(self, name: str, description: str, products: list[Product]):
-        """Инициализация категории. Все параметры обязательны."""
-        self.name = name
-        self.description = description
-        self.products = products
-
-        # Обновляем атрибуты класса
-        Category.total_categories += 1
-        Category.total_products += len(products)
-
-    def __repr__(self) -> str:
-        return f"Category('{self.name}', товаров: {len(self.products)})"
-
+from src.category import Category
+from src.product import Product
 
 def load_data_from_json(file_path: str) -> list[Category]:
     """Загружает данные из JSON и создает объекты Category и Product."""
